requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.0
anyio==4.4.0
attrs==24.2.0
babel==2.16.0
baml==0.19.1
baml-py==0.53.1
baml_core_ffi==0.1.7
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docxcompose==1.4.0
docxtpl==0.18.0
fastapi==0.112.1
filelock==3.15.4
Flask==3.0.3
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.25
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
MarkupSafe==2.1.5
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
ollama==0.1.9
openai==1.41.0
opentelemetry-api==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==23.2
platformdirs==4.3.4
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
PyJWT==2.3.0
pypandoc==1.13
pyparsing==2.4.7
pytest==7.4.4
pytest-asyncio==0.21.2
python-apt==2.4.0+ubuntu3
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.4
SecretStorage==3.3.1
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
tenacity==9.0.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typeguard==4.3.0
types-regex==2023.12.25.20240311
types-requests==2.32.0.20240712
typing_extensions==4.12.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.4
wadllib==1.3.6
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==1.0.0
